SVM bug fixes and improvements

+++ src/tangent/internal/base32.c
@@ -234,7 +234,7 @@ int base32_decode_character(char encoded, const char *alphabet) {
   const char *occurrence = strchr(alphabet, encoded);
 
   if (occurrence) {
-    return occurrence - alphabet;
+    return (int)(occurrence - alphabet);
   } else {
     return -1;
   }
